The rendering is refactored completely.
The markup is now create through a seperate hook function. This is more
dynamic and creates the possiblity to add the theme function better.

Todo: we need to test settings for a formatter and add a field instance to
the block. Settings need to be generic and also be able to be set on the
theme function.