Fix for class AddThis class loading. http://drupal.org/node/628486#comment-4987614
Some how when installing you cant uninstall the module and keep everything running. What happens
is that the install file is loaded and gives a error. It want to load hook_field_schema and
references the AddThis class. There fore we do a include_once for the class in AddThis.php.