Added alter hooks for markup and for the options, added service name reference.
I added two alter hooks. On is to alter the markup that is used to do stuff with
the render array. The other is to alter the options that are used to help in
rendering the markup for example entity, url and display settings.

I also added a #addthis_service to the render array that is support information
for alter implementations that want to do stuff with specific service elements.